Protección con Middleware y Policies

+++ app/Http/Controllers/PermissionController.php
@@ -2,61 +2,112 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\Permission\StoreRequest;
-use App\Http\Requests\Permission\UpdateRequest;
 use App\Permission;
 use App\Role;
+use App\Http\Requests\Permission\StoreRequest;
+use App\Http\Requests\Permission\UpdateRequest;
 use Illuminate\Http\Request;
 
 class PermissionController extends Controller
 {
-   
+    public function __construct()
+    {
+        $this->middleware('role:' . config('app.admin_role'));
+    }
+
+
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
     public function index()
     {
-        return view('theme.backoffice.pages.permission.index', [
-            'permissions'=>  Permission::all(),
+        $this->authorize('index',Role::class);
+        return view('theme.backoffice.pages.permission.index',[
+            'permissions' => Permission::all()
         ]);
     }
 
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
     public function create()
-    {   /* para pasar al SELECT de la Vista */
+    {
+        $this->authorize('create',Role::class);
         return view('theme.backoffice.pages.permission.create',[
-            'roles'=> Role::all(), 
+            'roles' => Role::all()
         ]);
     }
 
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
     public function store(StoreRequest $request, Permission $permission)
     {
-        $permission = $permission->store($request);        
+        $permission = $permission->store($request);
         return redirect()->route('backoffice.permission.show', $permission);
     }
-   
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Permission  $permission
+     * @return \Illuminate\Http\Response
+     */
     public function show(Permission $permission)
     {
-        return view('theme.backoffice.pages.permission.show', [
-            'permission'=>  $permission,
+        $this->authorize('view', $permission);
+        return view('theme.backoffice.pages.permission.show',[
+            'permission' => $permission
         ]);
     }
 
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  \App\Permission  $permission
+     * @return \Illuminate\Http\Response
+     */
     public function edit(Permission $permission)
     {
-        return view('theme.backoffice.pages.permission.edit', [
-            'permission'=>  $permission,
-            'roles'=> Role::all(), 
+        $this->authorize('update', $permission);
+        return view('theme.backoffice.pages.permission.edit',[
+            'permission' => $permission,
+            'roles' => Role::all()
         ]);
     }
 
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Permission  $permission
+     * @return \Illuminate\Http\Response
+     */
     public function update(UpdateRequest $request, Permission $permission)
     {
         $permission->my_update($request);
-        return redirect()->route('backoffice.permission.show', $permission);
+        return redirect()->route('backoffice.permission.show',$permission);
     }
-    
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  \App\Permission  $permission
+     * @return \Illuminate\Http\Response
+     */
     public function destroy(Permission $permission)
     {
+        $this->authorize('delete', $permission);
         $role = $permission->role;
         $permission->delete();
-        toast('Permiso Eliminado!','success', 'top-right');
+        toast('Permiso eliminado!','success', 'top-right');
         return redirect()->route('backoffice.role.show', $role);
     }
 }
